+++ migrations/versions/c09a21b13b0c_repository.py
@@ -30,4 +30,5 @@ def upgrade():
 
 
 def downgrade():
+    op.drop_table('repository')
     pass