+++ migrations/versions/f7c886c6084d_.py
@@ -5,9 +5,10 @@ Revises:
 Create Date: 2021-04-24 19:19:25.051290
 
 """
-from alembic import op
 import sqlalchemy as sa
+from alembic import op
 
+from src.models.user import ProviderType
 
 # revision identifiers, used by Alembic.
 revision = 'f7c886c6084d'
@@ -21,8 +22,10 @@ def upgrade():
                     sa.Column("id", sa.Integer, primary_key=True, autoincrement=False),
                     sa.Column("login", sa.String(100), nullable=False),
                     sa.Column("email", sa.String(100), nullable=True),
-                    sa.Column("twitter_username", sa.String(20), nullable=True))
+                    sa.Column("twitter_username", sa.String(20), nullable=True),
+                    sa.Column('provider', sa.Enum(ProviderType), nullable=False))
 
 
 def downgrade():
+    op.drop_table('user')
     pass