+++ src/models/__init__.py
@@ -1,2 +1,1 @@
 from src.models import repository, user
-