+++ src/res/repository.py
@@ -1,68 +1,116 @@
-import json
-from typing import Optional
+from typing import Optional, List
+
 from fastapi import APIRouter, Depends
-from fastapi.responses import JSONResponse
-from src.models.user import User as ModelUser
+from starlette import status
+from starlette.responses import JSONResponse
+
 from src.models.repository import Repository as ModelRepository
-from src.schemas.user import UserOutput
+from src.models.user import ProviderType, User as ModelUser
+from src.providers.dealer import Dealer
+from src.schemas.repository import Repository
+from src.schemas.user import UserOutputRepos
 from src.services.database import get_con, SESSION
-import requests
 
 router = APIRouter()
 
 
-@router.get('/{username}')
-def hi(username: str, from_local: Optional[bool] = False):
-    user = ModelUser.create_from_git(username)
+@router.get('/{username}', response_model=UserOutputRepos, status_code=status.HTTP_200_OK)
+def get_user(username: str, provider: ProviderType,
+             from_local: Optional[bool] = False, data_base: SESSION = Depends(get_con)):
+    if from_local:
+        user = ModelUser.get_user_by_username(username=username, db_session=data_base)
+        if not user:
+            return JSONResponse(status_code=status.HTTP_404_NOT_FOUND,
+                                content='User not found in database')
 
-    if not from_local:
-        repositories = ModelRepository.get_all_from_github(username)
-        repositories_names = []
+        repositories = ModelRepository.get_all_from_local(db_session=data_base, user_id=user.id)
+        names = [repository.name for repository in repositories]
+        return UserOutputRepos(id=user.id,
+                               login=user.login,
+                               email=user.email,
+                               twitter_username=user.twitter_username,
+                               repos=names)
 
-        for repository in repositories:
-            repositories_names.append(repository.name)
+    user = Dealer.get_user(username=username, provider=provider)
+    if not user:
+        return JSONResponse(status_code=status.HTTP_404_NOT_FOUND,
+                            content='User not found remotely')
 
-        return dict(id=user.id,
-                    login=user.login,
-                    twitter_username=user.twitter_username,
-                    email=user.email,
-                    repositories=repositories_names)
-    """
-    repositories = ModelRepository.get_all_from_local(user.id)
+    repositories = Dealer.get_all_repos(username=username, provider=provider)
+    names = [repository.name for repository in repositories]
 
-    repositories_names = []
-
-    for repository in repositories:
-        repositories_names.append(repository.name)
-
-    return dict(id=user.id,
-                login=user.login,
-                twitter_username=user.twitter_username,
-                email=user.email,
-                repositories=repositories_names)
-    """
-
-
-@router.get('/{username}/{reponame}')
-def get_repo(username: str, reponame: str, save_data: bool, db: SESSION = Depends(get_con)):
-    repo = ModelRepository.get_one_from_github(username=username, reponame=reponame)
-    if repo.user_id is not None:
-        try:
-            ModelUser.exists(db_session=db, index=repo.user_id)
-            return 'works'
-        except Exception as error:
-            print(error)
-        return f'is not none but don work, {repo.user_id}'
-    return 'Is none'
+    return UserOutputRepos(id=user.id,
+                           login=user.login,
+                           email=user.email,
+                           twitter_username=user.twitter_username,
+                           repos=names)
 
 
-@router.delete('/{id}')
-def delete_repo(index: int, db: SESSION = Depends(get_con)):
-    repo = ModelRepository.get_by_id(index=index, db_session=db)
-    if not repo:
-        return
-    status_user, message = repo.delete(connection=db)
-    if status_user:
-        return
-    return JSONResponse(status_code=500,
-                        content=dict(message='Database error', details=message))
+@router.get('/repos/{username}', status_code=status.HTTP_200_OK, response_model=List[Repository])
+def get_all_repositories(username: str, provider: ProviderType):
+    repositories = Dealer.get_all_repos(username=username,
+                                        provider=provider)
+    if not repositories:
+        return JSONResponse(status_code=status.HTTP_404_NOT_FOUND,
+                            content='User not found in remotely')
+    return repositories
+
+
+@router.get('/{username}/{repository_name}', status_code=status.HTTP_200_OK,
+            response_model=Repository)
+def get_repositories(username: str, repository_name: str, provider: ProviderType,
+                     save_data: bool, data_base: SESSION = Depends(get_con)):
+    repository = Dealer.get_repository(username=username,
+                                       repository_name=repository_name,
+                                       provider=provider)
+    if save_data:
+        user = Dealer.get_user(username=username, provider=provider)
+        if not ModelUser.get_user_by_id(db_session=data_base, index=user.id):
+            model_user = ModelUser(id=user.id,
+                                   login=user.login,
+                                   email=user.email,
+                                   twitter_username=user.twitter_username,
+                                   provider=provider)
+            status_user, message = model_user.save(connection=data_base)
+
+            if not status_user:
+                return JSONResponse(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+                                    content=dict(message='Database error', details=message))
+
+        if not ModelRepository.get_by_id(db_session=data_base, index=repository.id):
+            model_repository = ModelRepository(id=repository.id,
+                                               user_id=user.id,
+                                               name=repository.name,
+                                               private=repository.private,
+                                               created_at=repository.created_at,
+                                               updated_at=repository.updated_at,
+                                               size=repository.size,
+                                               stars=repository.stars,
+                                               watchers=repository.watchers)
+            status_repository, message = model_repository.save(connection=data_base)
+
+            if not status_repository:
+                return JSONResponse(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+                                    content=dict(message='Database error', details=message))
+    return repository
+
+
+@router.delete('delete/{username}')
+def delete_user(username: str, provider: ProviderType, data_base: SESSION = Depends(get_con)):
+    user = Dealer.get_user(username=username, provider=provider)
+    user = ModelUser.get_user_by_id(db_session=data_base, index=user.id)
+    if user:
+        user.delete(connection=data_base)
+    return 'User deleted'
+
+
+@router.delete('delete/{username}/{repository_name}')
+def delete_repo(username: str, repository_name: str,
+                provider: ProviderType, data_base: SESSION = Depends(get_con)):
+    repository = Dealer.get_repository(username=username,
+                                       repository_name=repository_name,
+                                       provider=provider)
+    repository = ModelRepository.get_by_id(db_session=data_base, index=repository.id)
+    if repository:
+        repository.delete(connection=data_base)
+    return 'Repository deleted'