+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# V 0.1.1
+Added:
+- GitHub, GitLab and Dealer classes
+
+Changed:
+- Database migrations
+- Models
+- Schemas
+- Routes
+
+Fixed:
+- Database bug related to login
 # V 0.0.1
 Added:
 - Migrations